split the LUT emitters
